+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: Option to completely skip (hide) a test step
+
 ## 1.4.0 - 2020-02-25
 
 * Added: New command Get-RPFilter to get filters